Implement PeraColors and themes
