--- composeTestApp/src/androidMain/kotlin/co/algorand/app/ui/theme/Theme.android.kt
@@ -1,19 +1,0 @@
-package co.algorand.app.ui.theme
-
-import android.app.Activity
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.platform.LocalView
-import androidx.core.view.WindowInsetsControllerCompat
-
-@Composable
-internal actual fun SystemAppearance(isDark: Boolean) {
-    val view = LocalView.current
-    LaunchedEffect(isDark) {
-        val window = (view.context as Activity).window
-        WindowInsetsControllerCompat(window, window.decorView).apply {
-            isAppearanceLightStatusBars = isDark
-            isAppearanceLightNavigationBars = isDark
-        }
-    }
-}
