+++ composeTestApp/src/commonMain/kotlin/co/algorand/app/App.kt
@@ -1,9 +1,9 @@
 package co.algorand.app
 
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import co.algorand.app.di.initKoinConfig
 import co.algorand.app.ui.navigation.AppNavigation
+import com.algorand.common.ui.theme.PeraTheme
 import org.jetbrains.compose.ui.tooling.preview.Preview
 import org.koin.compose.KoinApplication
 
@@ -13,7 +13,7 @@ internal fun App() {
     KoinApplication(
         application = initKoinConfig
     ) {
-        MaterialTheme {
+        PeraTheme {
             AppNavigation()
         }
     }
