+++ composeTestApp/src/commonMain/kotlin/co/algorand/app/ui/navigation/TopBar.kt
@@ -1,14 +1,13 @@
-package co.alogrand.app.ui.navigation
+package co.algorand.app.ui.navigation
 
+import algorand_android.composetestapp.generated.resources.Res
+import algorand_android.composetestapp.generated.resources.app_bar_header
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-import co.algorand.app.ui.theme.md_theme_light_primary
-import algorand_android.composetestapp.generated.resources.Res
-import algorand_android.composetestapp.generated.resources.app_bar_header
+import com.algorand.common.ui.theme.PeraTheme
 import org.jetbrains.compose.resources.stringResource
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -17,8 +16,8 @@ fun TopBar() {
     TopAppBar(
         colors =
         TopAppBarDefaults.topAppBarColors(
-            containerColor = md_theme_light_primary,
-            titleContentColor = Color.White,
+            containerColor = PeraTheme.colors.background,
+            titleContentColor = PeraTheme.colors.textMain,
         ),
         title = {
             Text(
