+++ composeTestApp/src/commonMain/kotlin/co/algorand/app/ui/screens/DiscoverScreen.kt
@@ -5,16 +5,15 @@ import algorand_android.composetestapp.generated.resources.nav_discover
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.navigation.NavController
 import co.algorand.app.ui.viewmodels.AlgorandBaseViewModel
 import co.algorand.app.ui.widgets.SuppressLint
+import com.algorand.common.ui.theme.PeraTheme
 import org.jetbrains.compose.resources.stringResource
 
 @SuppressLint("ComposableNaming")
@@ -28,12 +27,8 @@ fun DiscoverScreen(
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.SpaceEvenly,
-        modifier =
-        Modifier
-            .fillMaxWidth()
-            .fillMaxHeight()
-            .background(Color.White),
+        modifier = Modifier.fillMaxSize().background(PeraTheme.colors.background)
     ) {
-        Text(text = stringResource(Res.string.nav_discover))
+        Text(text = stringResource(Res.string.nav_discover), color = PeraTheme.colors.textMain)
     }
 }
