+++ composeTestApp/src/commonMain/kotlin/co/algorand/app/ui/screens/SettingsScreen.kt
@@ -6,11 +6,11 @@ import algorand_android.composetestapp.generated.resources.ic_light_mode
 import algorand_android.composetestapp.generated.resources.nav_settings
 import algorand_android.composetestapp.generated.resources.open_github
 import algorand_android.composetestapp.generated.resources.theme
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.widthIn
@@ -28,9 +28,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import co.algorand.app.openUrl
-import co.algorand.app.ui.theme.LocalThemeIsDark
+import co.algorand.app.ui.typography.PeraTypographyPreviewScreenNavigation
 import co.algorand.app.ui.viewmodels.AlgorandBaseViewModel
 import co.algorand.app.ui.widgets.SuppressLint
+import com.algorand.common.ui.theme.LocalThemeIsDark
+import com.algorand.common.ui.theme.PeraTheme
 import org.jetbrains.compose.resources.stringResource
 import org.jetbrains.compose.resources.vectorResource
 
@@ -45,10 +47,7 @@ fun SettingsScreen(
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
         verticalArrangement = Arrangement.SpaceEvenly,
-        modifier =
-        Modifier
-            .fillMaxWidth()
-            .fillMaxHeight()
+        modifier = Modifier.fillMaxSize().background(PeraTheme.colors.background)
     ) {
         Text(text = stringResource(Res.string.nav_settings))
