+++ wallet-sdk/src/iosMain/kotlin/com/algorand/common/ui/theme/Theme.ios.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2022 Pera Wallet, LDA
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License
+ */
+
+package com.algorand.common.ui.theme
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import platform.UIKit.UIApplication
+import platform.UIKit.UIStatusBarStyleDarkContent
+import platform.UIKit.UIStatusBarStyleLightContent
+import platform.UIKit.setStatusBarStyle
+
+@Composable
+internal actual fun SystemAppearance(isDark: Boolean) {
+    LaunchedEffect(isDark) {
+        UIApplication.sharedApplication.setStatusBarStyle(
+            if (isDark) UIStatusBarStyleDarkContent else UIStatusBarStyleLightContent
+        )
+    }
+}
